Fix float duration in API endpoints and use single version source
The previous duration fix only updated the model but missed the API
query parameters (still int). Also consolidate version to read from
pyproject.toml via importlib.metadata instead of hardcoding in
multiple places.

Bump version to 0.3.5.

+++ sipring/api/config.py
@@ -169,7 +169,7 @@ def _get_source_user(request: Request) -> Optional[str]:
 
 
 @router.post("/{id_or_slug}/test", response_model=RingResponse)
-async def test_config(id_or_slug: str, request: Request, duration: int = 3):
+async def test_config(id_or_slug: str, request: Request, duration: float = 3):
     """
     Test a ring configuration with a short ring.
